// apps/web/src/components/shared/form-builder/types.ts
export interface SelectOption {
  label: string;
  value: string | number;
}